# Conflicts:
#	TAmanagement/TAmanage/tests.py

+++ TAmanagement/TAmanage/commands.py
@@ -2,15 +2,14 @@ from .models import *
 import shlex
 
 
-
 class Commands:
     def __init__(self):
         self.cmdList = []
         self.addCmd(Role.Administrator, "Create Course", "/createCourse")
         self.addCmd(Role.Administrator, "Create User", "/createUser")
-        self.addCmd(Role.Administrator, "Assign TAs", "/assignTas")
+        self.addCmd(Role.Administrator, "Assign Instructors/TAs", "/assignTas")
         self.addCmd([Role.Administrator, Role.Instructor, Role.TA], "View/Edit Profile", "/viewProfile")
-        #self.addCmd([Role.Administrator, Role.Instructor, Role.TA], "Edit Profile", "/editProfile")
+        # self.addCmd([Role.Administrator, Role.Instructor, Role.TA], "Edit Profile", "/editProfile")
         self.addCmd([Role.Administrator, Role.Instructor, Role.TA], "List Courses", "/listCourses")
         self.addCmd([Role.Administrator, Role.Instructor, Role.TA], "List Users", "/listUsers")
 
@@ -46,11 +45,13 @@ class CommandWorker:
             'create': {
                 'course': self.create_course,
                 'user': self.create_user,
+                'lab': self.create_lab,
             },
             'edit': {
                 'course': self.edit_course,
                 'user': self.edit_user,
                 'profile': self.edit_profile,
+                'lab': self.edit_lab,
             },
             'list': {
                 'courses': self.list_courses,
@@ -59,11 +60,16 @@ class CommandWorker:
             'view': {
                 'profile': self.view_profile,
                 'user': self.view_user,
+                'labs': self.view_labs,
             },
-            'delete user':self.delete_user,
+            'delete': {
+                'user': self.delete_user,
+                'course': self.delete_course,
+            },
             'login': self.login,
             'logout': self.logout,
-            'assignTa' : self.assign_ta,
+            'assignTa': self.assign_ta,
+            'validate': self.validate,
         }
         while type(worker) is dict:
             try:
@@ -88,6 +94,56 @@ class CommandWorker:
         c.save()
         return 'Course Added'
 
+    def view_labs(self, cmd: [str]):
+        if not self.currentUser:
+            return 'Need to be logged in to list courses'
+        cs = Course.objects.get(name=cmd[0])
+        labs = Lab.objects.filter(course=cs)
+        return labs
+
+    def create_lab(self, cmd: [str]):
+        if not self.currentUser or not self.currentUser.has_role(Role.Administrator):
+            return 'Only an Administrator can create a lab'
+        if len(cmd) < 2:
+            return 'Invalid number of parameters'
+        if len(cmd) > 2:
+            return 'Invalid number of parameters'
+        cs = Course.objects.get(name=cmd[0])
+        c = Lab.objects.filter(course=cs, section=cmd[1])
+        if c:
+            return 'Lab Already Exists'
+        section = cmd[1]
+        c = Lab(course=cs, section=section)
+        c.save()
+        return "Lab Added"
+
+    def edit_lab(self, cmd: [str]):
+        if not self.currentUser or not self.currentUser.has_role(Role.Administrator):
+            return 'Only an Administrator can edit a lab'
+        if len(cmd) < 6:
+            return 'Invalid number of parameters'
+        if len(cmd) > 6:
+            return 'Invalid number of parameters'
+        c = Course.objects.get(name=cmd[0])
+        l = Lab.objects.get(course=c, section=cmd[1])
+        valid_dates = {"M", "T", "W", "R", "F", "S", "Online"}
+        if l:
+            l.section = cmd[1]
+            l.location = cmd[2]
+            l.startTime = cmd[3]
+            l.endTime = cmd[4]
+            l.save()
+            for ch in cmd[5]:
+                if valid_dates.intersection(ch):
+                    continue
+                else:
+                    return 'Invalid date(s)'
+            l.dates = cmd[5]
+        else:
+            return 'Lab does not yet exist'
+        l.save()
+        return 'Lab Updated'
+
     def list_courses(self, cmd: [str]):
         if not self.currentUser:
             return 'Need to be logged in to list courses'
@@ -104,14 +160,21 @@ class CommandWorker:
         users = User.objects.all()
         return users
 
-    def delete_user(self):
+    def delete_user(self, cmd: [str]):
         if not self.currentUser or not self.currentUser.has_role(Role.Administrator):
-            return "Only Admin can delete user"
-        
-        v=User.objects.get(self)
-        print(x)
+            return "Only Admin Can Delete User"
+        v = User.objects.get(self)
+        if v == self.currentUser:
+            return "Cannot Delete Yourself"
         v.delete()
         return "User Deleted"
+
+    def delete_course(self, cmd: [str]):
+        if not self.currentUser or not self.currentUser.has_role(Role.Administrator):
+            return "Only Admin Can Delete Course"
+        v = Course.objects.get(self)
+        v.delete()
+        return "Course Deleted"
 
     def view_profile(self, cmd: [str]):
         if not self.currentUser:
@@ -175,21 +238,23 @@ class CommandWorker:
         if u:
             u.firstName = cmd[1]
             u.lastName = cmd[2]
-            u.phone = cmd[3]
-            u.address = cmd[4]
-            u.officeHours = cmd[5]
-            u.officeHoursDates = cmd[6]
-            
-            for ch in cmd[6]:
+            u.role=cmd[3]
+            u.phone = cmd[4]
+            u.address = cmd[5]
+            u.officeHours = cmd[6]
+            u.officeHoursDates = cmd[7]
+
+            for ch in cmd[7]:
                 if valid_dates.intersection(ch):
                     continue
                 else:
                     return 'Invalid date(s)'
-            u.officeHoursDates = cmd[6]
-            u.officeLocation = cmd[7]
-            u.resume=cmd[8]
-            u.schedule=cmd[9]
-            u.preferences=cmd[10]
+            u.officeHoursDates = cmd[7]
+            u.officeLocation = cmd[8]
+            u.resume = cmd[9]
+            u.schedule = cmd[10]
+            u.preferences = cmd[11]
+            
         else:
             return 'User does not exist'
         u.save()
@@ -247,12 +312,12 @@ class CommandWorker:
             return 'Invalid number of parameters'
         courselist = Course.objects.all()
         for i in courselist:
-            talist = i.graderTAs
             for j in courselist:
                 # Check if there are any overlapping TAs for courses that occur over the same times.
-                if len(j.graderTAs.intersection(talist)) != 0 and i.name != j.name:
+                intersection = set(i.graderTAs.all()).intersection(set(j.graderTAs.all()))
+                if len(intersection) != 0 and i.name != j.name and not i.isOnline() and not j.isOnline():
                     if i.getStartTime() <= j.getStartTime() <= i.getEndTime() or i.getStartTime() <= j.getEndTime() <= i.getEndTime():
-                        return 'TA schedule conflict found'
+                        return 'Error: Overlapping TAs found in conflicting courses ' + i.name + ' and ' + j.name + '.'
                     elif j.getStartTime() <= i.getStartTime() <= j.getEndTime() or j.getStartTime() <= i.getEndTime() <= j.getEndTime():
-                        return 'TA schedule conflict found'
-        return 'Course is valid'
+                        return 'Error: Overlapping TAs found in conflicting courses ' + i.name + ' and ' + j.name + '.'
+        return 'TA assignments are valid'
